requirements.txt
# Base Django Stuff
Django==1.7.1
psycopg2==2.5.4
django-debug-toolbar==1.2.1
# DRF related
djangorestframework==2.4.3
django-filter==0.8
Markdown==2.4.1
# Flex Gallery
Pillow==2.4.0
django-fine-uploader
